+++ comp2013/src/main/java/comp2013/IMovable.java
@@ -1,13 +1,8 @@
 package comp2013;
 
 /**
- * 
- * @Project Snakee
- * @Description Fyrsta viðmótið
- * @Author Sigurður Sigurðardóttir
- * @version Ekki viss
- */ 
-
+ * Represents a movable snake object.
+ */
 public interface IMovable
 {
 	public int m_Direction = 0;